+++ Pod/Classes/Public/FOSRetrieveToOneRelationshipOperation.m
@@ -125,7 +125,7 @@
             if (blockSelf->_fetchRelatedEntityOp != nil) {
                 NSAssert(blockSelf.relationship.isOwnershipRelationship ||
                          !relDesc.isOptional ||
-                         [[[moc objectWithID:ownerId] class] forcePullForRelationship:relDesc givenJSON:[[moc objectWithID:ownerId] originalJson]] != FOSForcePullType_Never,
+                         [[moc objectWithID:ownerId] forcePullForRelationship:relDesc givenJSON:[[moc objectWithID:ownerId] originalJson]] != FOSForcePullType_Never,
                          @"Expected an ownership relationship, not a graph relationship.");
                 [blockSelf->_fetchRelatedEntityOp finishBinding];
                 childObj = [moc objectWithID:blockSelf->_fetchRelatedEntityOp.managedObjectID];
@@ -212,11 +212,11 @@
                     // too, so those owners are not the parent object that performed the pull.
                     NSAssert([owner primitiveValueForKey:relDesc.name] == nil ||
                              [owner primitiveValueForKey:relDesc.name] == childObj ||
-                             [[owner class] forcePullForRelationship:relDesc givenJSON:owner.originalJson] == FOSForcePullType_Always,
+                             [owner forcePullForRelationship:relDesc givenJSON:owner.originalJson] == FOSForcePullType_Always,
                              @"Relationship already bound???");
 
                     // Set the forward relationship
-                    FOSForcePullType forcePull = [[owner class] forcePullForRelationship:relDesc givenJSON: owner.originalJson];
+                    FOSForcePullType forcePull = [owner forcePullForRelationship:relDesc givenJSON: owner.originalJson];
                     
                     if ([owner primitiveValueForKey:relDesc.name] == nil ||
                         forcePull == FOSForcePullType_Always) {